pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "self_hosted_search_engine"
version = "0.1.0"
description = "Self-hosted deep research agent"
authors = [{name = "Akeem"}]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4==4.12.3",
    "chromadb==0.5.3",
    "duckdb==1.0.0",
    "flask==3.0.3",
    "flask-cors==4.0.0",
    "httpx==0.27.0",
    "langchain-core==0.2.11",
    "langdetect==1.0.9",
    "langgraph==0.0.58",
    "langsmith==0.1.147",
    "lxml==5.3.0",
    "numpy==1.26.4",
    "opentelemetry-sdk==1.37.0",
    "pdfminer.six==20231228",
    "playwright==1.47.0",
    "prometheus-client==0.20.0",
    "pydantic==2.9.2",
    "pytest==8.3.3",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "scrapy==2.11.2",
    "scrapy-playwright==0.0.39",
    "tiktoken==0.7.0",
    "trafilatura==1.7.0",
    "watchdog==4.0.1",
    "whoosh==2.7.4",
]

[tool.setuptools]
packages = [
    "backend",
    "crawler",
    "engine",
    "llm",
    "rank",
    "search",
    "server",
    "seed_loader",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.mypy]
python_version = "3.11"
pretty = true
show_error_context = true
ignore_missing_imports = true
implicit_reexport = true
warn_unreachable = false
warn_return_any = false
warn_unused_ignores = false
disable_error_code = [
    "attr-defined",
    "call-overload",
    "index",
    "misc",
    "operator",
    "return-value",
    "assignment",
    "arg-type",
    "no-redef",
    "valid-type",
]

[[tool.mypy.overrides]]
module = "backend.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "crawler.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "engine.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "llm.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "rank.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "search.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "server.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "seed_loader.*"
ignore_errors = true
